A lot of new things and fixed some bugs

+++ test/performance/helper.go
@@ -1,7 +1,6 @@
 package performance
 
 import (
-	"log"
 	"os"
 	"path/filepath"
 	"testing"
@@ -10,6 +9,7 @@ import (
 	"github.com/natsukagami/kjudge/db"
 	"github.com/natsukagami/kjudge/models"
 	"github.com/natsukagami/kjudge/server/auth"
+	"github.com/natsukagami/kjudge/test/performance/suites"
 	"github.com/natsukagami/kjudge/worker"
 	"github.com/natsukagami/kjudge/worker/queue"
 	"github.com/natsukagami/kjudge/worker/sandbox"
@@ -18,23 +18,20 @@ import (
 
 type BenchmarkContext struct {
 	tdir     string
+	remove   bool
 	db       *db.DB
 	user     *models.User
 	contest  *models.Contest
 	problems map[string]*models.Problem
 }
 
-func NewBenchmarkContext(tmpDir string) (*BenchmarkContext, error) {
+func NewBenchmarkContext(tmpDir string, remove bool) (*BenchmarkContext, error) {
 	benchDB, err := db.New(filepath.Join(tmpDir, "bench.db"))
 	if err != nil {
-		err2 := os.RemoveAll(tmpDir)
-		if err2 != nil {
-			return nil, errors.Wrapf(err2, "while handling %v", err)
-		}
 		return nil, err
 	}
 
-	ctx := &BenchmarkContext{tdir: tmpDir, db: benchDB, problems: make(map[string]*models.Problem)}
+	ctx := &BenchmarkContext{tdir: tmpDir, remove: remove, db: benchDB, problems: make(map[string]*models.Problem)}
 
 	if err := ctx.writeContest(); err != nil {
 		return nil, err
@@ -48,11 +45,14 @@ func NewBenchmarkContext(tmpDir string) (*BenchmarkContext, error) {
 }
 
 func (ctx *BenchmarkContext) Close() error {
-	if err := os.RemoveAll(ctx.tdir); err != nil {
-		return err
-	}
 	if err := ctx.db.Close(); err != nil {
 		return err
+	}
+
+	if ctx.remove {
+		if err := os.RemoveAll(ctx.tdir); err != nil {
+			return err
+		}
 	}
 	return nil
 }
@@ -83,7 +83,7 @@ func (ctx *BenchmarkContext) writeUser() error {
 	return errors.Wrap(ctx.user.Write(ctx.db), "creating user")
 }
 
-func (ctx *BenchmarkContext) writeProblem(testset *PerfTestSet) error {
+func (ctx *BenchmarkContext) writeProblem(testset *suites.PerfTestSet) error {
 	problem, err := testset.AddToDB(ctx.db, 2403, len(ctx.problems)+1, ctx.contest.ID)
 	if err != nil {
 		return errors.Wrapf(err, "creating testset %v's problem", testset.Name)
@@ -96,45 +96,39 @@ func (ctx *BenchmarkContext) writeProblem(testset *PerfTestSet) error {
 const testSolution = `#include "solution.hpp"
 `
 
-func (ctx *BenchmarkContext) writeSolutions(N int, problemName string) error {
+func (ctx *BenchmarkContext) writeSolution(problemName string) error {
 	problem := ctx.problems[problemName]
-	for i := 0; i < N; i++ {
-		sub := models.Submission{
-			ProblemID:   problem.ID,
-			UserID:      ctx.user.ID,
-			Source:      []byte(testSolution),
-			Language:    models.LanguageCpp,
-			SubmittedAt: time.Now(),
-			Verdict:     models.VerdictIsInQueue,
-		}
-		if err := sub.Write(ctx.db); err != nil {
-			return err
-		}
+	sub := models.Submission{
+		ProblemID:   problem.ID,
+		UserID:      ctx.user.ID,
+		Source:      []byte(testSolution),
+		Language:    models.LanguageCpp,
+		SubmittedAt: time.Now(),
+		Verdict:     models.VerdictIsInQueue,
+	}
+	if err := sub.Write(ctx.db); err != nil {
+		return err
+	}
 
-		job := models.NewJobScore(sub.ID)
-		if err := job.Write(ctx.db); err != nil {
-			return err
-		}
+	job := models.NewJobScore(sub.ID)
+	if err := job.Write(ctx.db); err != nil {
+		return err
 	}
 	return nil
 }
 
-func RunSingleTest(b *testing.B, ctx *BenchmarkContext, testset *PerfTestSet, sandboxName string) {
-	log.Printf("running %v %v %v times", testset.Name, sandboxName, b.N)
+func RunSingleTest(b *testing.B, ctx *BenchmarkContext, testset *suites.PerfTestSet, sandboxName string) {
 	sandbox, err := worker.NewSandbox(
 		sandboxName,
 		sandbox.IgnoreWarnings(true),
 		sandbox.EnableSandboxLogs(false))
-
 	if err != nil {
 		b.Fatal(err)
 	}
 
-	log.Printf("Generating %v solutions", b.N)
 	for i := 0; i < b.N; i++ {
-		ctx.writeSolutions(b.N, testset.Name)
+		ctx.writeSolution(testset.Name)
 	}
-	log.Printf("Generated solutions")
 
 	queue := queue.NewQueue(ctx.db, sandbox, queue.CompileLogs(false), queue.RunLogs(false), queue.ScoreLogs(false))
 
@@ -145,10 +139,9 @@ func RunSingleTest(b *testing.B, ctx *BenchmarkContext, testset *PerfTestSet, sa
 	if err := ctx.assertRunComplete(testset); err != nil {
 		b.Fatal(err)
 	}
-	log.Printf("Finished running queue")
 }
 
 // TODO
-func (ctx *BenchmarkContext) assertRunComplete(testset *PerfTestSet) error {
+func (ctx *BenchmarkContext) assertRunComplete(testset *suites.PerfTestSet) error {
 	return nil
 }
